tripletex/requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform>=1.70
requests
python-dotenv
httpx
pytest
